# DialogueClients takes advantage of the EndpointChannel bind optimization