@svc-autorelease 2020-06-04
Released [1.56.0](https://github.com/palantir/dialogue/releases/tag/1.56.0)